refactor: rename AdminDashboard folder to admin-dashboard
Adopt kebab-case naming convention for the Platform Hub page directory.
