+++ frontend/web/components/pages/admin-dashboard/components/InstanceMetricsCards.tsx
