+++ frontend/web/components/pages/admin-dashboard/components/IntegrationAdoptionTable.tsx
