+++ frontend/web/components/pages/admin-dashboard/components/ReleasePipelineStatsTable.tsx
