+++ frontend/web/components/pages/admin-dashboard/components/StaleFlagsTable.tsx
