+++ frontend/web/components/pages/admin-dashboard/components/UsageTrendsChart.tsx
